Updated naming conventions and formatting.
Updated and commented code in pockets.hpp. renamed pockets_unit_test to pockets_methods_unit_test.hpp. added formatting to README.md and fixed implicit conversions in wraptext.hpp

+++ wraptext.hpp
@@ -8,7 +8,7 @@
 #define wraptext_hpp
 
 inline sf::String wrapText(sf::String string, unsigned width, const sf::Font &font, unsigned charicterSize, bool bold = false){
-    unsigned currentOffset = 0;
+    float currentOffset = 0;
     bool firstWord = true;
     std::size_t wordBegining = 0;
